requirements.txt
openai
anthropic
google-generativeai
tavily-python
python-dotenv
streamlit
gTTS
pydantic
tenacity
pydantic-settings
elevenlabs
fpdf
python-docx
openpyxl
requests
boto3
edge-tts
nest_asyncio
